requirements.txt
# Core document generation
python-docx>=1.1.0
docxtpl>=0.17.0
Pillow>=10.0.0

# Excel and data handling
pandas>=2.2.0
openpyxl>=3.1.2

# PDF conversion options
docx2pdf>=0.1.8         # used on Windows
pypandoc>=1.13          # fallback converter if needed

# GUI
tkinterweb>=3.12.0; platform_system != "Windows"  # optional, helps with Tkinter rendering
tqdm>=4.67.1            # progress helper

# Compatibility and packaging
importlib-resources>=6.1.0
